+++ DatadogCore/Tests/Datadog/Mocks/DatadogInternal/DatadogCoreProxy.swift
@@ -46,6 +46,7 @@ internal class DatadogCoreProxy: DatadogCoreProtocol {
             encryption: nil,
             contextProvider: DatadogContextProvider(context: context),
             applicationVersion: context.version,
+            maxBatchesPerUpload: .mockRandom(min: 1, max: 100),
             backgroundTasksEnabled: .mockAny()
         )