+++ DatadogCore/Tests/Datadog/RUM/RUMFeatureTests.swift
@@ -71,6 +71,7 @@ class RUMFeatureTests: XCTestCase {
                 )
             ),
             applicationVersion: randomApplicationVersion,
+            maxBatchesPerUpload: .mockRandom(min: 1, max: 100),
             backgroundTasksEnabled: randomBackgroundTasksEnabled
         )
         defer { core.flushAndTearDown() }
@@ -138,6 +139,7 @@ class RUMFeatureTests: XCTestCase {
             encryption: nil,
             contextProvider: .mockAny(),
             applicationVersion: .mockAny(),
+            maxBatchesPerUpload: .mockRandom(min: 1, max: 100),
             backgroundTasksEnabled: .mockAny()
         )
         defer { core.flushAndTearDown() }