+++ DatadogRUM/Tests/Mocks/RUMDataModelMocks.swift
@@ -109,7 +109,7 @@ extension RUMOperatingSystem: RandomMockable {
 
 extension RUMViewEvent.DD.Configuration: RandomMockable {
     public static func mockRandom() -> RUMViewEvent.DD.Configuration {
-        return .init(sessionReplaySampleRate: .mockRandom(min: 0, max: 100), sessionSampleRate: .mockRandom(min: 0, max: 100))
+        return .init(sessionReplaySampleRate: .mockRandom(min: 0, max: 100), sessionSampleRate: .mockRandom(min: 0, max: 100), startSessionReplayRecordingManually: nil)
     }
 }
 
@@ -479,9 +479,11 @@ extension TelemetryConfigurationEvent: RandomMockable {
                     actionNameAttribute: nil,
                     allowFallbackToLocalStorage: nil,
                     allowUntrustedEvents: nil,
+                    backgroundTasksEnabled: .mockAny(),
+                    batchProcessingLevel: .mockAny(),
                     batchSize: .mockAny(),
-                    batchUploadFrequency: .mockAny(),
-                    defaultPrivacyLevel: .mockAny(),
+                    batchUploadFrequency: .mockRandom(),
+                    defaultPrivacyLevel: .mockRandom(),
                     forwardConsoleLogs: nil,
                     forwardErrorsToLogs: nil,
                     forwardReports: nil,